feat(l2): ethrex l2 --dev (#3836)
**Motivation**
We want to be able to initialize the full l2 by running `ethrex l2
--dev`
<!-- Why does this pull request exist? What are its goals? -->

**Description**
Move all the logic for initializing the l2 to `cmd/ethrex`
Now calling `ethrex l2 --dev` does the following

- Remove L1 and L2 DBs
- Initialize an L1
- Deploy contracts
- Initialize an L2

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------
