--- .github/workflows/pr-main_l2_contracts.yaml
@@ -1,35 +1,0 @@
-name: L2 Contracts
-on:
-  push:
-    branches: ["main"]
-  pull_request:
-    branches: ["**"]
-    paths:
-      - "crates/l2/**"
-      - ".github/workflows/pr-main_l2_contracts.yaml"
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
-  cancel-in-progress: true
-
-jobs:
-  test:
-    # "Test" is a required check, don't change the name
-    name: Test
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout sources
-        uses: actions/checkout@v4
-      - name: Setup Rust Environment
-        uses: ./.github/actions/setup-rust
-
-      - name: Install solc
-        uses: pontem-network/get-solc@master
-        with:
-          version: v0.8.29
-          token: ${{ secrets.GITHUB_TOKEN || '' }}
-
-      - name: Run test of deployer.rs
-        run: |
-          cd crates/l2/contracts
-          cargo test
