+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
   "crates/common/trie",
   "crates/common/crypto",
   "crates/l2/",
-  "crates/l2/contracts",
   "crates/l2/common",
   "crates/l2/prover",
   "crates/l2/prover/zkvm/interface",
