+++ Dockerfile
@@ -39,6 +39,8 @@ ARG BUILD_FLAGS=""
 COPY crates ./crates
 COPY cmd ./cmd
 COPY metrics ./metrics
+COPY tooling ./tooling
+COPY fixtures/genesis ./fixtures/genesis
 COPY Cargo.* ./
 RUN cargo build --release $BUILD_FLAGS
