+++ cmd/ethrex/l2/initializers.rs
@@ -131,7 +131,7 @@ fn init_metrics(opts: &L1Options, tracker: TaskTracker) {
     tracker.spawn(metrics_api);
 }
 
-fn init_tracing(opts: &L2Options) {
+pub fn init_tracing(opts: &L2Options) {
     if !opts.sequencer_opts.no_monitor {
         let level_filter = EnvFilter::builder()
             .parse_lossy("debug,tower_http::trace=debug,reqwest_tracing=off,hyper=off,libsql=off,ethrex::initializers=off,ethrex::l2::initializers=off,ethrex::l2::command=off");
@@ -150,8 +150,6 @@ pub async fn init_l2(opts: L2Options) -> eyre::Result<()> {
     if opts.node_opts.evm == EvmEngine::REVM {
         panic!("L2 Doesn't support REVM, use LEVM instead.");
     }
-
-    init_tracing(&opts);
 
     let data_dir = set_datadir(&opts.node_opts.datadir);
     let rollup_store_dir = data_dir.clone() + "/rollup_store";
