+++ cmd/ethrex/l2/mod.rs
@@ -1,10 +1,11 @@
 mod initializers;
 
 pub mod command;
+pub mod deployer;
 pub mod options;
 
 pub use command::L2Command;
-pub use initializers::init_l2;
+pub use initializers::{init_l2, init_tracing};
 pub use options::{
     BlockProducerOptions, CommitterOptions, EthOptions, Options as L2Options,
     ProofCoordinatorOptions, SequencerOptions, WatcherOptions,
