--- crates/l2/contracts/Cargo.toml
@@ -1,52 +1,0 @@
-[package]
-name = "ethrex_l2-contracts"
-version = "0.1.0"
-edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.38.0", features = ["full"] }
-ethereum-types = { workspace = true, features = ["serialize"] }
-bytes = { version = "1.6.0", features = ["serde"] }
-secp256k1.workspace = true
-keccak-hash.workspace = true
-hex = "0.4.3"
-serde_json = "1.0.138"
-colored = "2.1.0"
-lazy_static = "1.5.0"
-tracing.workspace = true
-tracing-subscriber = { version = "0.3", features = ["fmt"] }
-thiserror.workspace = true
-clap.workspace = true
-clap_complete.workspace = true
-eyre.workspace = true
-reqwest.workspace = true
-
-ethrex-l2 = { path = "../../l2" }
-ethrex-sdk = { path = "../../l2/sdk" }
-ethrex-common = { path = "../../common" }
-ethrex-l2-common = { path = "../common" }
-ethrex-l2-rpc.workspace = true
-ethrex-rpc = { path = "../../networking/rpc" }
-genesis-tool = { path = "../../../tooling/genesis" }
-
-[[bin]]
-name = "ethrex_l2_l1_deployer"
-path = "bin/deployer/main.rs"
-
-[[bin]]
-name = "ethrex_l2_system_contracts_updater"
-path = "bin/system_contracts_updater/main.rs"
-
-[lints.clippy]
-unwrap_used = "deny"
-expect_used = "deny"
-indexing_slicing = "deny"
-as_conversions = "deny"
-unnecessary_cast = "warn"
-panic = "deny"
-
-[build-dependencies]
-ethrex-sdk = { path = "../../l2/sdk" }
-hex.workspace = true
