+++ crates/l2/docker-compose-l2-store.overrides.yaml
@@ -11,8 +11,6 @@ services:
       --authrpc.port 8552 
       --datadir /store 
       --proof-coordinator.addr 0.0.0.0
-      --l1.bridge-address ${ETHREX_WATCHER_BRIDGE_ADDRESS}
-      --l1.on-chain-proposer-address ${ETHREX_COMMITTER_ON_CHAIN_PROPOSER_ADDRESS}
       --block-producer.coinbase-address 0x0007a881CD95B1484fca47615B64803dad620C8d
       --committer.l1-private-key 0x385c546456b6a603a1cfcaa9ec9494ba4832da08dd6bcf4de9a71e4a01b74924
       --proof-coordinator.l1-private-key 0x39725efee3fb28614de3bacaffe4cc4bd8c436257e2c8bb887c4b5c4be45e76d
