+++ crates/l2/sdk/contract_utils/src/compile.rs
@@ -16,6 +16,7 @@ pub fn compile_contract(
     contract_path: &Path,
     runtime_bin: bool,
     remappings: Option<&[(&str, PathBuf)]>,
+    allow_paths: &[&Path],
 ) -> Result<(), ContractCompilationError> {
     let bin_flag = if runtime_bin {
         "--bin-runtime"
@@ -41,13 +42,11 @@ pub fn compile_contract(
             .to_str()
             .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
     )
-    .arg("--overwrite")
-    .arg("--allow-paths")
-    .arg(
-        output_dir
-            .to_str()
-            .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-    );
+    .arg("--overwrite");
+
+    if !allow_paths.is_empty() {
+        apply_allow_paths(&mut cmd, allow_paths)?;
+    }
 
     let cmd_succeeded = cmd
         .spawn()
@@ -81,5 +80,22 @@ fn apply_remappings(
             cmd.arg(format!("{prefix}={path_str}"));
         }
     }
+    Ok(())
+}
+
+fn apply_allow_paths(
+    cmd: &mut Command,
+    allow_paths: &[&Path],
+) -> Result<(), ContractCompilationError> {
+    cmd.arg("--allow-paths");
+    let joined_paths = allow_paths
+        .iter()
+        .map(|p| {
+            p.to_str()
+                .ok_or(ContractCompilationError::FailedToGetStringFromPath)
+        })
+        .collect::<Result<Vec<_>, _>>()?
+        .join(",");
+    cmd.arg(joined_paths);
     Ok(())
 }
